app/requirements.txt
streamlit
google.cloud.logging
google-cloud-translate